requirements.txt
dropbox
selenium<4
click
pyyaml
ConfigParser
PyPyDispatcher
sentry-sdk
cmutils
chromedriver-autoinstaller
psutil
pytz
gideon-api
python-dateutil
gspread
oauth2client
google-api-python-client
google-auth-httplib2
google-auth-oauthlib